requirements.txt
Flask
tensorflow
scikit-learn
numpy
pandas
matplotlib
Pillow
joblib


